+++ src/main/java/com/example/teamprojeect/controller/work/WorkFileController.java
@@ -1,16 +1,16 @@
 package com.example.teamprojeect.controller.work;
 
-
 import com.example.teamprojeect.service.WorkService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
+@Controller
 @RequiredArgsConstructor
-@RestController
-@RequestMapping("/like/work/*")
-public class WorkLikeController {
+@RequestMapping("/file/artist/*")
+public class WorkFileController {
+    // 필드
     private final WorkService workService;
 }
