+++ src/main/java/com/example/teamprojeect/domain/dao/work/WorkDAO.java
@@ -21,8 +21,8 @@ public class WorkDAO {
     };
 
     // 작품 수정 신청 및 관리자 수정 승인 (매개변수로 usreNumber를 받아서 사용자라면 수정신청을, 관리자라면 바로 수정을 실행)
-    public int modifyApply(@Param("workVO") WorkVO workVO,  @Param("userNumber") Long userNumber) {
-        return workMapper.updateApply(workVO, userNumber);
+    public boolean modifyApply(@Param("workVO") WorkVO workVO,  @Param("userNumber") Long userNumber) {
+        return workMapper.updateApply(workVO, userNumber) == 1;
     }
 
     // 작품 삭제
