--- src/main/java/com/example/teamprojeect/service/ReplyService.java
@@ -1,13 +1,0 @@
-package com.example.teamprojeect.service;
-
-import com.example.teamprojeect.domain.dao.artist.ArtistReplyDAO;
-import com.example.teamprojeect.domain.dao.show.ShowReplyDAO;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ReplyService {
-    private final ArtistReplyDAO artistReplyDAO;
-    private final ShowReplyDAO showReplyDAO;
-}
